requirements.txt
torch>=2.0
transformers>=4.40.0
trl>=0.14.0
datasets>=2.0
accelerate>=0.30.0
pyyaml>=6.0
